@@ -40,11 +40,13 @@ setup(
         "Intended Audience :: Developers",
         "Topic :: Scientific/Engineering :: Artificial Intelligence",
         "License :: OSI Approved :: Apache Software License",
-        "Programming Language :: Python :: 3.6",
-        "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
     ],
-    packages=find_packages(exclude=["*.tests", "*.tests.*", "tests.*", "tests"]),
+    packages=find_packages(
+        exclude=["*.tests", "*.tests.*", "tests.*", "tests", "colabs", "*.ipynb"]
+    ),
     zip_safe=False,
     install_requires=[
         "cloudpickle",
@@ -53,7 +55,12 @@ setup(
         "Pillow>=4.2.1",
         "protobuf>=3.6",
         "pyyaml>=3.1.0",
+        "gym>=0.21.0",
+        "pettingzoo==1.15.0",
+        "numpy==1.21.2",
+        "filelock>=3.4.0",
     ],
-    python_requires=">=3.6.1",
-    cmdclass={"verify": VerifyVersionCommand},
+    python_requires=">=3.8.13,<=3.10.8",
+    # TODO: Remove this once mypy stops having spurious setuptools issues.
+    cmdclass={"verify": VerifyVersionCommand},  # type: ignore
 )
