@@ -3,7 +3,7 @@ import sys
 
 from setuptools import setup, find_packages
 from setuptools.command.install import install
-from mlagents.plugins import ML_AGENTS_STATS_WRITER
+from mlagents.plugins import ML_AGENTS_STATS_WRITER, ML_AGENTS_TRAINER_TYPE
 import mlagents.trainers
 
 VERSION = mlagents.trainers.__version__
@@ -48,9 +48,9 @@ setup(
         "Intended Audience :: Developers",
         "Topic :: Scientific/Engineering :: Artificial Intelligence",
         "License :: OSI Approved :: Apache Software License",
-        "Programming Language :: Python :: 3.6",
-        "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
     ],
     # find_namespace_packages will recurse through the directories and find all the packages
     packages=find_packages(exclude=["*.tests", "*.tests.*", "tests.*", "tests"]),
@@ -65,10 +65,10 @@ setup(
         "protobuf>=3.6",
         "pyyaml>=3.1.0",
         # Windows ver. of PyTorch doesn't work from PyPi. Installation:
-        # https://github.com/Unity-Technologies/ml-agents/blob/release_19_docs/docs/Installation.md#windows-installing-pytorch
+        # https://github.com/Unity-Technologies/ml-agents/blob/release_20_docs/docs/Installation.md#windows-installing-pytorch
         # Torch only working on python 3.9 for 1.8.0 and above. Details see:
         # https://github.com/pytorch/pytorch/issues/50014
-        "torch>=1.8.0,<1.9.0;(platform_system!='Windows' and python_version>='3.9')",
+        "torch>=1.8.0,<=1.11.0;(platform_system!='Windows' and python_version>='3.9')",
         "torch>=1.6.0,<1.9.0;(platform_system!='Windows' and python_version<'3.9')",
         "tensorboard>=1.15",
         # cattrs 1.1.0 dropped support for python 3.6, but 1.0.0 doesn't work for python 3.9
@@ -77,9 +77,9 @@ setup(
         "cattrs>=1.1.0,<1.7; python_version>='3.8'",
         "attrs>=19.3.0",
         'pypiwin32==223;platform_system=="Windows"',
-        "importlib_metadata; python_version<'3.8'",
+        "importlib_metadata==4.4; python_version<'3.8'",
     ],
-    python_requires=">=3.6.1",
+    python_requires=">=3.8.13,<=3.10.8",
     entry_points={
         "console_scripts": [
             "mlagents-learn=mlagents.trainers.learn:main",
@@ -89,6 +89,10 @@ setup(
         ML_AGENTS_STATS_WRITER: [
             "default=mlagents.plugins.stats_writer:get_default_stats_writers"
         ],
+        ML_AGENTS_TRAINER_TYPE: [
+            "default=mlagents.plugins.trainer_type:get_default_trainer_types"
+        ],
     },
-    cmdclass={"verify": VerifyVersionCommand},
+    # TODO: Remove this once mypy stops having spurious setuptools issues.
+    cmdclass={"verify": VerifyVersionCommand},  # type: ignore
 )
