@@ -76,8 +76,9 @@ namespace Unity.MLAgentsExamples
         private ModelOverrider m_ModelOverrider;
 
         private const float k_RewardMultiplier = 0.01f;
-        void Awake()
+        protected override void Awake()
         {
+            base.Awake();
             Board = GetComponent<Match3Board>();
             m_ModelOverrider = GetComponent<ModelOverrider>();
         }
