@@ -44,8 +44,9 @@ public class PushBlockWithInputAgentBasic : Agent
 
     EnvironmentParameters m_ResetParams;
 
-    void Awake()
+    protected override void Awake()
     {
+        base.Awake();
         m_PushBlockSettings = FindObjectOfType<PushBlockWithInputSettings>();
 
         goalDetect = block.GetComponent<GoalDetectWithInput>();
