@@ -1,4 +1,7 @@
-#if UNITY_EDITOR || UNITY_STANDALONE_WIN || UNITY_STANDALONE_OSX || UNITY_STANDALONE_LINUX
+#if UNITY_EDITOR || UNITY_STANDALONE
+#define MLA_SUPPORTED_TRAINING_PLATFORM
+#endif
+#if MLA_SUPPORTED_TRAINING_PLATFORM
 // <auto-generated>
 //     Generated by the protocol buffer compiler.  DO NOT EDIT!
 //     source: mlagents_envs/communicator_objects/unity_to_external.proto
