@@ -4,12 +4,15 @@ import os
 import uuid
 import shutil
 import glob
+
 import yaml
 import hashlib
 
 from zipfile import ZipFile
 from sys import platform
 from typing import Tuple, Optional, Dict, Any
+
+from filelock import FileLock
 
 from mlagents_envs.env_utils import validate_environment_path
 
@@ -21,34 +24,40 @@ logger = get_logger(__name__)
 BLOCK_SIZE = 8192
 
 
-def get_local_binary_path(name: str, url: str) -> str:
+def get_local_binary_path(name: str, url: str, tmp_dir: Optional[str] = None) -> str:
     """
     Returns the path to the executable previously downloaded with the name argument. If
     None is found, the executable at the url argument will be downloaded and stored
     under name for future uses.
     :param name: The name that will be given to the folder containing the extracted data
     :param url: The URL of the zip file
+    :param: tmp_dir: Optional override for the temporary directory to save binaries and zips in.
     """
     NUMBER_ATTEMPTS = 5
-    path = get_local_binary_path_if_exists(name, url)
-    if path is None:
-        logger.debug(
-            f"Local environment {name} not found, downloading environment from {url}"
-        )
-    for attempt in range(NUMBER_ATTEMPTS):  # Perform 5 attempts at downloading the file
-        if path is not None:
-            break
-        try:
-            download_and_extract_zip(url, name)
-        except Exception:
-            if attempt + 1 < NUMBER_ATTEMPTS:
-                logger.warning(
-                    f"Attempt {attempt + 1} / {NUMBER_ATTEMPTS}"
-                    ": Failed to download and extract binary."
-                )
-            else:
-                raise
-        path = get_local_binary_path_if_exists(name, url)
+    tmp_dir = tmp_dir or tempfile.gettempdir()
+    lock = FileLock(os.path.join(tmp_dir, name + ".lock"))
+    with lock:
+        path = get_local_binary_path_if_exists(name, url, tmp_dir=tmp_dir)
+        if path is None:
+            logger.debug(
+                f"Local environment {name} not found, downloading environment from {url}"
+            )
+        for attempt in range(
+            NUMBER_ATTEMPTS
+        ):  # Perform 5 attempts at downloading the file
+            if path is not None:
+                break
+            try:
+                download_and_extract_zip(url, name, tmp_dir=tmp_dir)
+            except Exception:
+                if attempt + 1 < NUMBER_ATTEMPTS:
+                    logger.warning(
+                        f"Attempt {attempt + 1} / {NUMBER_ATTEMPTS}"
+                        ": Failed to download and extract binary."
+                    )
+                else:
+                    raise
+            path = get_local_binary_path_if_exists(name, url, tmp_dir=tmp_dir)
 
     if path is None:
         raise FileNotFoundError(
@@ -58,15 +67,16 @@ def get_local_binary_path(name: str, url: str) -> str:
     return path
 
 
-def get_local_binary_path_if_exists(name: str, url: str) -> Optional[str]:
+def get_local_binary_path_if_exists(name: str, url: str, tmp_dir: str) -> Optional[str]:
     """
     Recursively searches for a Unity executable in the extracted files folders. This is
     platform dependent : It will only return a Unity executable compatible with the
     computer's OS. If no executable is found, None will be returned.
     :param name: The name/identifier of the executable
     :param url: The url the executable was downloaded from (for verification)
+    :param: tmp_dir: Optional override for the temporary directory to save binaries and zips in.
     """
-    _, bin_dir = get_tmp_dir()
+    _, bin_dir = get_tmp_dirs(tmp_dir)
     extension = None
 
     if platform == "linux" or platform == "linux2":
@@ -94,40 +104,54 @@ def get_local_binary_path_if_exists(name: str, url: str) -> Optional[str]:
         return None
 
 
-def get_tmp_dir() -> Tuple[str, str]:
+def _get_tmp_dir_helper(tmp_dir: Optional[str] = None) -> Tuple[str, str]:
+    tmp_dir = tmp_dir or ("/tmp" if platform == "darwin" else tempfile.gettempdir())
+    MLAGENTS = "ml-agents-binaries"
+    TMP_FOLDER_NAME = "tmp"
+    BINARY_FOLDER_NAME = "binaries"
+    mla_directory = os.path.join(tmp_dir, MLAGENTS)
+    if not os.path.exists(mla_directory):
+        os.makedirs(mla_directory)
+        os.chmod(mla_directory, 16877)
+    zip_directory = os.path.join(tmp_dir, MLAGENTS, TMP_FOLDER_NAME)
+    if not os.path.exists(zip_directory):
+        os.makedirs(zip_directory)
+        os.chmod(zip_directory, 16877)
+    bin_directory = os.path.join(tmp_dir, MLAGENTS, BINARY_FOLDER_NAME)
+    if not os.path.exists(bin_directory):
+        os.makedirs(bin_directory)
+        os.chmod(bin_directory, 16877)
+    return zip_directory, bin_directory
+
+
+def get_tmp_dirs(tmp_dir: Optional[str] = None) -> Tuple[str, str]:
     """
     Returns the path to the folder containing the downloaded zip files and the extracted
     binaries. If these folders do not exist, they will be created.
     :retrun: Tuple containing path to : (zip folder, extracted files folder)
     """
-    TEMPDIR = "/tmp" if platform == "darwin" else tempfile.gettempdir()
-    MLAGENTS = "ml-agents-binaries"
-    TMP_FOLDER_NAME = "tmp"
-    BINARY_FOLDER_NAME = "binaries"
-    mla_directory = os.path.join(TEMPDIR, MLAGENTS)
-    if not os.path.exists(mla_directory):
-        os.makedirs(mla_directory)
-        os.chmod(mla_directory, 16877)
-    zip_directory = os.path.join(TEMPDIR, MLAGENTS, TMP_FOLDER_NAME)
-    if not os.path.exists(zip_directory):
-        os.makedirs(zip_directory)
-        os.chmod(zip_directory, 16877)
-    bin_directory = os.path.join(TEMPDIR, MLAGENTS, BINARY_FOLDER_NAME)
-    if not os.path.exists(bin_directory):
-        os.makedirs(bin_directory)
-        os.chmod(bin_directory, 16877)
-    return (zip_directory, bin_directory)
+    # TODO: Once we don't use python 3.7 we should just use exists_ok=True when creating the dirs to avoid this.
+    # Should only be able to error out 3 times (once for each subdir).
+    for _attempt in range(3):
+        try:
+            return _get_tmp_dir_helper(tmp_dir)
+        except FileExistsError:
+            continue
+    return _get_tmp_dir_helper(tmp_dir)
 
 
-def download_and_extract_zip(url: str, name: str) -> None:
+def download_and_extract_zip(
+    url: str, name: str, tmp_dir: Optional[str] = None
+) -> None:
     """
     Downloads a zip file under a URL, extracts its contents into a folder with the name
     argument and gives chmod 755 to all the files it contains. Files are downloaded and
     extracted into special folders in the temp folder of the machine.
     :param url: The URL of the zip file
     :param name: The name that will be given to the folder containing the extracted data
+    :param: tmp_dir: Optional override for the temporary directory to save binaries and zips in.
     """
-    zip_dir, bin_dir = get_tmp_dir()
+    zip_dir, bin_dir = get_tmp_dirs(tmp_dir)
     url_hash = "-" + hashlib.md5(url.encode()).hexdigest()
     binary_path = os.path.join(bin_dir, name + url_hash)
     if os.path.exists(binary_path):
@@ -137,7 +161,7 @@ def download_and_extract_zip(url: str, name: str) -> None:
     try:
         request = urllib.request.urlopen(url, timeout=30)
     except urllib.error.HTTPError as e:  # type: ignore
-        e.msg += " " + url
+        e.reason = f"{e.reason} {url}"
         raise
     zip_size = int(request.headers["content-length"])
     zip_file_path = os.path.join(zip_dir, str(uuid.uuid4()) + ".zip")
@@ -189,11 +213,11 @@ def load_remote_manifest(url: str) -> Dict[str, Any]:
     """
     Converts a remote yaml file into a Python dictionary
     """
-    tmp_dir, _ = get_tmp_dir()
+    tmp_dir, _ = get_tmp_dirs()
     try:
         request = urllib.request.urlopen(url, timeout=30)
     except urllib.error.HTTPError as e:  # type: ignore
-        e.msg += " " + url
+        e.reason = f"{e.reason} {url}"
         raise
     manifest_path = os.path.join(tmp_dir, str(uuid.uuid4()) + ".yaml")
     with open(manifest_path, "wb") as manifest:
