@@ -16,6 +16,7 @@ class RemoteRegistryEntry(BaseRegistryEntry):
         darwin_url: Optional[str],
         win_url: Optional[str],
         additional_args: Optional[List[str]] = None,
+        tmp_dir: Optional[str] = None,
     ):
         """
         A RemoteRegistryEntry is an implementation of BaseRegistryEntry that uses a
@@ -39,6 +40,7 @@ class RemoteRegistryEntry(BaseRegistryEntry):
         self._darwin_url = darwin_url
         self._win_url = win_url
         self._add_args = additional_args
+        self._tmp_dir_override = tmp_dir
 
     def make(self, **kwargs: Any) -> BaseEnv:
         """
@@ -58,7 +60,9 @@ class RemoteRegistryEntry(BaseRegistryEntry):
                 f"The entry {self.identifier} does not contain a valid url for this "
                 "platform"
             )
-        path = get_local_binary_path(self.identifier, url)
+        path = get_local_binary_path(
+            self.identifier, url, tmp_dir=self._tmp_dir_override
+        )
         if "file_name" in kwargs:
             kwargs.pop("file_name")
         args: List[str] = []
