@@ -42,11 +42,11 @@ class BehaviorIdentifiers(NamedTuple):
 
 def create_name_behavior_id(name: str, team_id: int) -> str:
     """
-   Reconstructs fully qualified behavior name from name and team_id
-   :param name: brain name
-   :param team_id: team ID
-   :return: name_behavior_id
-   """
+    Reconstructs fully qualified behavior name from name and team_id
+    :param name: brain name
+    :param team_id: team ID
+    :return: name_behavior_id
+    """
     return name + "?team=" + str(team_id)
 
 
