@@ -8,7 +8,7 @@ from mlagents.trainers.model_saver.model_saver import BaseModelSaver
 from mlagents.trainers.settings import TrainerSettings, SerializationSettings
 from mlagents.trainers.policy.torch_policy import TorchPolicy
 from mlagents.trainers.optimizer.torch_optimizer import TorchOptimizer
-from mlagents.trainers.torch.model_serialization import ModelSerializer
+from mlagents.trainers.torch_entities.model_serialization import ModelSerializer
 
 
 logger = get_logger(__name__)
