@@ -5,8 +5,10 @@ from collections import defaultdict
 
 from mlagents.trainers.buffer import AgentBuffer, AgentBufferField
 from mlagents.trainers.trajectory import ObsUtil
-from mlagents.trainers.torch.components.bc.module import BCModule
-from mlagents.trainers.torch.components.reward_providers import create_reward_provider
+from mlagents.trainers.torch_entities.components.bc.module import BCModule
+from mlagents.trainers.torch_entities.components.reward_providers import (
+    create_reward_provider,
+)
 
 from mlagents.trainers.policy.torch_policy import TorchPolicy
 from mlagents.trainers.optimizer import Optimizer
@@ -15,7 +17,7 @@ from mlagents.trainers.settings import (
     RewardSignalSettings,
     RewardSignalType,
 )
-from mlagents.trainers.torch.utils import ModelUtils
+from mlagents.trainers.torch_entities.utils import ModelUtils
 
 
 class TorchOptimizer(Optimizer):
@@ -129,6 +131,12 @@ class TorchOptimizer(Optimizer):
         }
         next_mem = _mem
         return all_value_tensors, all_next_memories, next_mem
+
+    def update_reward_signals(self, batch: AgentBuffer) -> Dict[str, float]:
+        update_stats: Dict[str, float] = {}
+        for reward_provider in self.reward_signals.values():
+            update_stats.update(reward_provider.update(batch))
+        return update_stats
 
     def get_trajectory_value_estimates(
         self,
