@@ -3,7 +3,7 @@
 # Contains an implementation of MA-POCA.
 
 from collections import defaultdict
-from typing import cast, Dict
+from typing import cast, Dict, Union, Any, Type
 
 import numpy as np
 
@@ -11,18 +11,23 @@ from mlagents_envs.side_channel.stats_side_channel import StatsAggregationMethod
 from mlagents_envs.logging_util import get_logger
 from mlagents_envs.base_env import BehaviorSpec
 from mlagents.trainers.buffer import BufferKey, RewardSignalUtil
-from mlagents.trainers.trainer.rl_trainer import RLTrainer
+from mlagents.trainers.trainer.on_policy_trainer import OnPolicyTrainer
+from mlagents.trainers.trainer.trainer_utils import lambda_return
 from mlagents.trainers.policy import Policy
 from mlagents.trainers.policy.torch_policy import TorchPolicy
-from mlagents.trainers.poca.optimizer_torch import TorchPOCAOptimizer
+from mlagents.trainers.poca.optimizer_torch import TorchPOCAOptimizer, POCASettings
 from mlagents.trainers.trajectory import Trajectory
 from mlagents.trainers.behavior_id_utils import BehaviorIdentifiers
-from mlagents.trainers.settings import TrainerSettings, POCASettings
+from mlagents.trainers.settings import TrainerSettings
+
+from mlagents.trainers.torch_entities.networks import SimpleActor, SharedActorCritic
 
 logger = get_logger(__name__)
 
+TRAINER_NAME = "poca"
 
-class POCATrainer(RLTrainer):
+
+class POCATrainer(OnPolicyTrainer):
     """The POCATrainer is an implementation of the MA-POCA algorithm."""
 
     def __init__(
@@ -47,17 +52,19 @@ class POCATrainer(RLTrainer):
         """
         super().__init__(
             behavior_name,
+            reward_buff_cap,
             trainer_settings,
             training,
             load,
+            seed,
             artifact_path,
-            reward_buff_cap,
         )
         self.hyperparameters: POCASettings = cast(
             POCASettings, self.trainer_settings.hyperparameters
         )
         self.seed = seed
         self.policy: TorchPolicy = None  # type: ignore
+        self.optimizer: TorchPOCAOptimizer = None  # type: ignore
         self.collected_group_rewards: Dict[str, int] = defaultdict(lambda: 0)
 
     def _process_trajectory(self, trajectory: Trajectory) -> None:
@@ -72,7 +79,7 @@ class POCATrainer(RLTrainer):
         agent_buffer_trajectory = trajectory.to_agentbuffer()
         # Update the normalization
         if self.is_training:
-            self.policy.update_normalization(agent_buffer_trajectory)
+            self.policy.actor.update_normalization(agent_buffer_trajectory)
             self.optimizer.critic.update_normalization(agent_buffer_trajectory)
 
         # Get all value estimates
@@ -193,56 +200,6 @@ class POCATrainer(RLTrainer):
         size_of_buffer = self.update_buffer.num_experiences
         return size_of_buffer > self.hyperparameters.buffer_size
 
-    def _update_policy(self):
-        """
-        Uses demonstration_buffer to update the policy.
-        The reward signal generators must be updated in this method at their own pace.
-        """
-        buffer_length = self.update_buffer.num_experiences
-        self.cumulative_returns_since_policy_update.clear()
-
-        # Make sure batch_size is a multiple of sequence length. During training, we
-        # will need to reshape the data into a batch_size x sequence_length tensor.
-        batch_size = (
-            self.hyperparameters.batch_size
-            - self.hyperparameters.batch_size % self.policy.sequence_length
-        )
-        # Make sure there is at least one sequence
-        batch_size = max(batch_size, self.policy.sequence_length)
-
-        n_sequences = max(
-            int(self.hyperparameters.batch_size / self.policy.sequence_length), 1
-        )
-
-        advantages = np.array(
-            self.update_buffer[BufferKey.ADVANTAGES].get_batch(), dtype=np.float32
-        )
-        self.update_buffer[BufferKey.ADVANTAGES].set(
-            (advantages - advantages.mean()) / (advantages.std() + 1e-10)
-        )
-        num_epoch = self.hyperparameters.num_epoch
-        batch_update_stats = defaultdict(list)
-        for _ in range(num_epoch):
-            self.update_buffer.shuffle(sequence_length=self.policy.sequence_length)
-            buffer = self.update_buffer
-            max_num_batch = buffer_length // batch_size
-            for i in range(0, max_num_batch * batch_size, batch_size):
-                update_stats = self.optimizer.update(
-                    buffer.make_mini_batch(i, i + batch_size), n_sequences
-                )
-                for stat_name, value in update_stats.items():
-                    batch_update_stats[stat_name].append(value)
-
-        for stat, stat_list in batch_update_stats.items():
-            self._stats_reporter.add_stat(stat, np.mean(stat_list))
-
-        if self.optimizer.bc_module:
-            update_stats = self.optimizer.bc_module.update()
-            for stat, val in update_stats.items():
-                self._stats_reporter.add_stat(stat, val)
-        self._clear_update_buffer()
-        return True
-
     def end_episode(self) -> None:
         """
         A signal that the Episode has ended. The buffer must be reset.
@@ -252,7 +209,7 @@ class POCATrainer(RLTrainer):
         super().end_episode()
         self.collected_group_rewards.clear()
 
-    def create_torch_policy(
+    def create_policy(
         self, parsed_behavior_id: BehaviorIdentifiers, behavior_spec: BehaviorSpec
     ) -> TorchPolicy:
         """
@@ -261,40 +218,23 @@ class POCATrainer(RLTrainer):
         :param behavior_spec: specifications for policy construction
         :return policy
         """
+        actor_cls: Union[Type[SimpleActor], Type[SharedActorCritic]] = SimpleActor
+        actor_kwargs: Dict[str, Any] = {
+            "conditional_sigma": False,
+            "tanh_squash": False,
+        }
+
         policy = TorchPolicy(
             self.seed,
             behavior_spec,
-            self.trainer_settings,
-            condition_sigma_on_obs=False,  # Faster training for POCA
-            separate_critic=True,  # Match network architecture with TF
+            self.trainer_settings.network_settings,
+            actor_cls,
+            actor_kwargs,
         )
         return policy
 
-    def create_poca_optimizer(self) -> TorchPOCAOptimizer:
+    def create_optimizer(self) -> TorchPOCAOptimizer:
         return TorchPOCAOptimizer(self.policy, self.trainer_settings)
-
-    def add_policy(
-        self, parsed_behavior_id: BehaviorIdentifiers, policy: Policy
-    ) -> None:
-        """
-        Adds policy to trainer.
-        :param parsed_behavior_id: Behavior identifiers that the policy should belong to.
-        :param policy: Policy to associate with name_behavior_id.
-        """
-        if not isinstance(policy, TorchPolicy):
-            raise RuntimeError(f"policy {policy} must be an instance of TorchPolicy.")
-        self.policy = policy
-        self.policies[parsed_behavior_id.behavior_id] = policy
-        self.optimizer = self.create_poca_optimizer()
-        for _reward_signal in self.optimizer.reward_signals.keys():
-            self.collected_rewards[_reward_signal] = defaultdict(lambda: 0)
-
-        self.model_saver.register(self.policy)
-        self.model_saver.register(self.optimizer)
-        self.model_saver.initialize_or_load()
-
-        # Needed to resume loads properly
-        self._step = policy.get_current_step()
 
     def get_policy(self, name_behavior_id: str) -> Policy:
         """
@@ -304,14 +244,6 @@ class POCATrainer(RLTrainer):
 
         return self.policy
 
-
-def lambda_return(r, value_estimates, gamma=0.99, lambd=0.8, value_next=0.0):
-    returns = np.zeros_like(r)
-    returns[-1] = r[-1] + gamma * value_next
-    for t in reversed(range(0, r.size - 1)):
-        returns[t] = (
-            gamma * lambd * returns[t + 1]
-            + r[t]
-            + (1 - lambd) * gamma * value_estimates[t + 1]
-        )
-    return returns
+    @staticmethod
+    def get_trainer_name() -> str:
+        return TRAINER_NAME
