@@ -1,4 +1,4 @@
-from typing import Any, Dict, List, Tuple, Optional
+from typing import Any, Dict, List
 import numpy as np
 from mlagents.torch_utils import torch, default_device
 import copy
@@ -9,13 +9,10 @@ from mlagents.trainers.policy import Policy
 from mlagents_envs.base_env import DecisionSteps, BehaviorSpec
 from mlagents_envs.timers import timed
 
-from mlagents.trainers.settings import TrainerSettings
-from mlagents.trainers.torch.networks import SimpleActor, SharedActorCritic, GlobalSteps
+from mlagents.trainers.settings import NetworkSettings
+from mlagents.trainers.torch_entities.networks import GlobalSteps
 
-from mlagents.trainers.torch.utils import ModelUtils
-from mlagents.trainers.buffer import AgentBuffer
-from mlagents.trainers.torch.agent_action import AgentAction
-from mlagents.trainers.torch.action_log_probs import ActionLogProbs
+from mlagents.trainers.torch_entities.utils import ModelUtils
 
 EPSILON = 1e-7  # Small value to avoid divide by zero
 
@@ -25,10 +22,9 @@ class TorchPolicy(Policy):
         self,
         seed: int,
         behavior_spec: BehaviorSpec,
-        trainer_settings: TrainerSettings,
-        tanh_squash: bool = False,
-        separate_critic: bool = True,
-        condition_sigma_on_obs: bool = True,
+        network_settings: NetworkSettings,
+        actor_cls: type,
+        actor_kwargs: Dict[str, Any],
     ):
         """
         Policy that uses a multilayer perceptron to map the observations to actions. Could
@@ -36,46 +32,26 @@ class TorchPolicy(Policy):
         continuous actions, as well as recurrent networks.
         :param seed: Random seed.
         :param behavior_spec: Assigned BehaviorSpec object.
-        :param trainer_settings: Defined training parameters.
-        :param load: Whether a pre-trained model will be loaded or a new one created.
-        :param tanh_squash: Whether to use a tanh function on the continuous output,
-        or a clipped output.
+        :param network_settings: Defined network parameters.
+        :param actor_cls: The type of Actor
+        :param actor_kwargs: Keyword args for the Actor class
         """
-        super().__init__(
-            seed, behavior_spec, trainer_settings, tanh_squash, condition_sigma_on_obs
-        )
+        super().__init__(seed, behavior_spec, network_settings)
         self.global_step = (
             GlobalSteps()
         )  # could be much simpler if TorchPolicy is nn.Module
-        self.grads = None
 
         self.stats_name_to_update_name = {
             "Losses/Value Loss": "value_loss",
             "Losses/Policy Loss": "policy_loss",
         }
-        if separate_critic:
-            self.actor = SimpleActor(
-                observation_specs=self.behavior_spec.observation_specs,
-                network_settings=trainer_settings.network_settings,
-                action_spec=behavior_spec.action_spec,
-                conditional_sigma=self.condition_sigma_on_obs,
-                tanh_squash=tanh_squash,
-            )
-            self.shared_critic = False
-        else:
-            reward_signal_configs = trainer_settings.reward_signals
-            reward_signal_names = [
-                key.value for key, _ in reward_signal_configs.items()
-            ]
-            self.actor = SharedActorCritic(
-                observation_specs=self.behavior_spec.observation_specs,
-                network_settings=trainer_settings.network_settings,
-                action_spec=behavior_spec.action_spec,
-                stream_names=reward_signal_names,
-                conditional_sigma=self.condition_sigma_on_obs,
-                tanh_squash=tanh_squash,
-            )
-            self.shared_critic = True
+
+        self.actor = actor_cls(
+            observation_specs=self.behavior_spec.observation_specs,
+            network_settings=network_settings,
+            action_spec=behavior_spec.action_spec,
+            **actor_kwargs,
+        )
 
         # Save the m_size needed for export
         self._export_m_size = self.m_size
@@ -83,7 +59,6 @@ class TorchPolicy(Policy):
         self.m_size = self.actor.memory_size
 
         self.actor.to(default_device())
-        self._clip_action = not tanh_squash
 
     @property
     def export_memory_size(self) -> int:
@@ -104,49 +79,6 @@ class TorchPolicy(Policy):
                 )
         return mask
 
-    def update_normalization(self, buffer: AgentBuffer) -> None:
-        """
-        If this policy normalizes vector observations, this will update the norm values in the graph.
-        :param buffer: The buffer with the observations to add to the running estimate
-        of the distribution.
-        """
-
-        if self.normalize:
-            self.actor.update_normalization(buffer)
-
-    @timed
-    def sample_actions(
-        self,
-        obs: List[torch.Tensor],
-        masks: Optional[torch.Tensor] = None,
-        memories: Optional[torch.Tensor] = None,
-        seq_len: int = 1,
-    ) -> Tuple[AgentAction, ActionLogProbs, torch.Tensor, torch.Tensor]:
-        """
-        :param obs: List of observations.
-        :param masks: Loss masks for RNN, else None.
-        :param memories: Input memories when using RNN, else None.
-        :param seq_len: Sequence length when using RNN.
-        :return: Tuple of AgentAction, ActionLogProbs, entropies, and output memories.
-        """
-        actions, log_probs, entropies, memories = self.actor.get_action_and_stats(
-            obs, masks, memories, seq_len
-        )
-        return (actions, log_probs, entropies, memories)
-
-    def evaluate_actions(
-        self,
-        obs: List[torch.Tensor],
-        actions: AgentAction,
-        masks: Optional[torch.Tensor] = None,
-        memories: Optional[torch.Tensor] = None,
-        seq_len: int = 1,
-    ) -> Tuple[ActionLogProbs, torch.Tensor]:
-        log_probs, entropies = self.actor.get_stats(
-            obs, actions, masks, memories, seq_len
-        )
-        return log_probs, entropies
-
     @timed
     def evaluate(
         self, decision_requests: DecisionSteps, global_agent_ids: List[str]
@@ -164,21 +96,15 @@ class TorchPolicy(Policy):
         memories = torch.as_tensor(self.retrieve_memories(global_agent_ids)).unsqueeze(
             0
         )
-
-        run_out = {}
         with torch.no_grad():
-            action, log_probs, entropy, memories = self.sample_actions(
+            action, run_out, memories = self.actor.get_action_and_stats(
                 tensor_obs, masks=masks, memories=memories
             )
-        action_tuple = action.to_action_tuple()
-        run_out["action"] = action_tuple
-        # This is the clipped action which is not saved to the buffer
-        # but is exclusively sent to the environment.
-        env_action_tuple = action.to_action_tuple(clip=self._clip_action)
-        run_out["env_action"] = env_action_tuple
-        run_out["log_probs"] = log_probs.to_log_probs_tuple()
-        run_out["entropy"] = ModelUtils.to_numpy(entropy)
-        run_out["learning_rate"] = 0.0
+        run_out["action"] = action.to_action_tuple()
+        if "log_probs" in run_out:
+            run_out["log_probs"] = run_out["log_probs"].to_log_probs_tuple()
+        if "entropy" in run_out:
+            run_out["entropy"] = ModelUtils.to_numpy(run_out["entropy"])
         if self.use_recurrent:
             run_out["memory_out"] = ModelUtils.to_numpy(memories).squeeze(0)
         return run_out
