@@ -2,26 +2,31 @@
 # ## ML-Agent Learning (PPO)
 # Contains an implementation of PPO as described in: https://arxiv.org/abs/1707.06347
 
-from collections import defaultdict
-from typing import cast
+from typing import cast, Type, Union, Dict, Any
 
 import numpy as np
 
-from mlagents_envs.logging_util import get_logger
 from mlagents_envs.base_env import BehaviorSpec
+from mlagents_envs.logging_util import get_logger
 from mlagents.trainers.buffer import BufferKey, RewardSignalUtil
-from mlagents.trainers.trainer.rl_trainer import RLTrainer
-from mlagents.trainers.policy import Policy
+from mlagents.trainers.trainer.on_policy_trainer import OnPolicyTrainer
+from mlagents.trainers.policy.policy import Policy
+from mlagents.trainers.trainer.trainer_utils import get_gae
+from mlagents.trainers.optimizer.torch_optimizer import TorchOptimizer
 from mlagents.trainers.policy.torch_policy import TorchPolicy
-from mlagents.trainers.ppo.optimizer_torch import TorchPPOOptimizer
+from mlagents.trainers.ppo.optimizer_torch import TorchPPOOptimizer, PPOSettings
 from mlagents.trainers.trajectory import Trajectory
 from mlagents.trainers.behavior_id_utils import BehaviorIdentifiers
-from mlagents.trainers.settings import TrainerSettings, PPOSettings
+from mlagents.trainers.settings import TrainerSettings
+
+from mlagents.trainers.torch_entities.networks import SimpleActor, SharedActorCritic
 
 logger = get_logger(__name__)
 
+TRAINER_NAME = "ppo"
 
-class PPOTrainer(RLTrainer):
+
+class PPOTrainer(OnPolicyTrainer):
     """The PPOTrainer is an implementation of the PPO algorithm."""
 
     def __init__(
@@ -46,17 +51,19 @@ class PPOTrainer(RLTrainer):
         """
         super().__init__(
             behavior_name,
+            reward_buff_cap,
             trainer_settings,
             training,
             load,
+            seed,
             artifact_path,
-            reward_buff_cap,
         )
         self.hyperparameters: PPOSettings = cast(
             PPOSettings, self.trainer_settings.hyperparameters
         )
         self.seed = seed
-        self.policy: Policy = None  # type: ignore
+        self.shared_critic = self.hyperparameters.shared_critic
+        self.policy: TorchPolicy = None  # type: ignore
 
     def _process_trajectory(self, trajectory: Trajectory) -> None:
         """
@@ -73,7 +80,7 @@ class PPOTrainer(RLTrainer):
 
         # Update the normalization
         if self.is_training:
-            self.policy.update_normalization(agent_buffer_trajectory)
+            self.policy.actor.update_normalization(agent_buffer_trajectory)
             self.optimizer.critic.update_normalization(agent_buffer_trajectory)
 
         # Get all value estimates
@@ -157,65 +164,12 @@ class PPOTrainer(RLTrainer):
         if trajectory.done_reached:
             self._update_end_episode_stats(agent_id, self.optimizer)
 
-    def _is_ready_update(self):
-        """
-        Returns whether or not the trainer has enough elements to run update model
-        :return: A boolean corresponding to whether or not update_model() can be run
-        """
-        size_of_buffer = self.update_buffer.num_experiences
-        return size_of_buffer > self.hyperparameters.buffer_size
+    def create_optimizer(self) -> TorchOptimizer:
+        return TorchPPOOptimizer(  # type: ignore
+            cast(TorchPolicy, self.policy), self.trainer_settings  # type: ignore
+        )  # type: ignore
 
-    def _update_policy(self):
-        """
-        Uses demonstration_buffer to update the policy.
-        The reward signal generators must be updated in this method at their own pace.
-        """
-        buffer_length = self.update_buffer.num_experiences
-        self.cumulative_returns_since_policy_update.clear()
-
-        # Make sure batch_size is a multiple of sequence length. During training, we
-        # will need to reshape the data into a batch_size x sequence_length tensor.
-        batch_size = (
-            self.hyperparameters.batch_size
-            - self.hyperparameters.batch_size % self.policy.sequence_length
-        )
-        # Make sure there is at least one sequence
-        batch_size = max(batch_size, self.policy.sequence_length)
-
-        n_sequences = max(
-            int(self.hyperparameters.batch_size / self.policy.sequence_length), 1
-        )
-
-        advantages = np.array(
-            self.update_buffer[BufferKey.ADVANTAGES].get_batch(), dtype=np.float32
-        )
-        self.update_buffer[BufferKey.ADVANTAGES].set(
-            (advantages - advantages.mean()) / (advantages.std() + 1e-10)
-        )
-        num_epoch = self.hyperparameters.num_epoch
-        batch_update_stats = defaultdict(list)
-        for _ in range(num_epoch):
-            self.update_buffer.shuffle(sequence_length=self.policy.sequence_length)
-            buffer = self.update_buffer
-            max_num_batch = buffer_length // batch_size
-            for i in range(0, max_num_batch * batch_size, batch_size):
-                update_stats = self.optimizer.update(
-                    buffer.make_mini_batch(i, i + batch_size), n_sequences
-                )
-                for stat_name, value in update_stats.items():
-                    batch_update_stats[stat_name].append(value)
-
-        for stat, stat_list in batch_update_stats.items():
-            self._stats_reporter.add_stat(stat, np.mean(stat_list))
-
-        if self.optimizer.bc_module:
-            update_stats = self.optimizer.bc_module.update()
-            for stat, val in update_stats.items():
-                self._stats_reporter.add_stat(stat, val)
-        self._clear_update_buffer()
-        return True
-
-    def create_torch_policy(
+    def create_policy(
         self, parsed_behavior_id: BehaviorIdentifiers, behavior_spec: BehaviorSpec
     ) -> TorchPolicy:
         """
@@ -224,48 +178,27 @@ class PPOTrainer(RLTrainer):
         :param behavior_spec: specifications for policy construction
         :return policy
         """
+        actor_cls: Union[Type[SimpleActor], Type[SharedActorCritic]] = SimpleActor
+        actor_kwargs: Dict[str, Any] = {
+            "conditional_sigma": False,
+            "tanh_squash": False,
+        }
+        if self.shared_critic:
+            reward_signal_configs = self.trainer_settings.reward_signals
+            reward_signal_names = [
+                key.value for key, _ in reward_signal_configs.items()
+            ]
+            actor_cls = SharedActorCritic
+            actor_kwargs.update({"stream_names": reward_signal_names})
+
         policy = TorchPolicy(
             self.seed,
             behavior_spec,
-            self.trainer_settings,
-            condition_sigma_on_obs=False,  # Faster training for PPO
-            separate_critic=True,  # Match network architecture with TF
+            self.trainer_settings.network_settings,
+            actor_cls,
+            actor_kwargs,
         )
         return policy
-
-    def create_ppo_optimizer(self) -> TorchPPOOptimizer:
-        return TorchPPOOptimizer(  # type: ignore
-            cast(TorchPolicy, self.policy), self.trainer_settings  # type: ignore
-        )  # type: ignore
-
-    def add_policy(
-        self, parsed_behavior_id: BehaviorIdentifiers, policy: Policy
-    ) -> None:
-        """
-        Adds policy to trainer.
-        :param parsed_behavior_id: Behavior identifiers that the policy should belong to.
-        :param policy: Policy to associate with name_behavior_id.
-        """
-        if self.policy:
-            logger.warning(
-                "Your environment contains multiple teams, but {} doesn't support adversarial games. Enable self-play to \
-                    train adversarial games.".format(
-                    self.__class__.__name__
-                )
-            )
-        self.policy = policy
-        self.policies[parsed_behavior_id.behavior_id] = policy
-
-        self.optimizer = self.create_ppo_optimizer()
-        for _reward_signal in self.optimizer.reward_signals.keys():
-            self.collected_rewards[_reward_signal] = defaultdict(lambda: 0)
-
-        self.model_saver.register(self.policy)
-        self.model_saver.register(self.optimizer)
-        self.model_saver.initialize_or_load()
-
-        # Needed to resume loads properly
-        self._step = policy.get_current_step()
 
     def get_policy(self, name_behavior_id: str) -> Policy:
         """
@@ -275,34 +208,6 @@ class PPOTrainer(RLTrainer):
 
         return self.policy
 
-
-def discount_rewards(r, gamma=0.99, value_next=0.0):
-    """
-    Computes discounted sum of future rewards for use in updating value estimate.
-    :param r: List of rewards.
-    :param gamma: Discount factor.
-    :param value_next: T+1 value estimate for returns calculation.
-    :return: discounted sum of future rewards as list.
-    """
-    discounted_r = np.zeros_like(r)
-    running_add = value_next
-    for t in reversed(range(0, r.size)):
-        running_add = running_add * gamma + r[t]
-        discounted_r[t] = running_add
-    return discounted_r
-
-
-def get_gae(rewards, value_estimates, value_next=0.0, gamma=0.99, lambd=0.95):
-    """
-    Computes generalized advantage estimate for use in updating policy.
-    :param rewards: list of rewards for time-steps t to T.
-    :param value_next: Value estimate for time-step T+1.
-    :param value_estimates: list of value estimates for time-steps t to T.
-    :param gamma: Discount factor.
-    :param lambd: GAE weighing factor.
-    :return: list of advantage estimates for time-steps t to T.
-    """
-    value_estimates = np.append(value_estimates, value_next)
-    delta_t = rewards + gamma * value_estimates[1:] - value_estimates[:-1]
-    advantage = discount_rewards(r=delta_t, gamma=gamma * lambd)
-    return advantage
+    @staticmethod
+    def get_trainer_name() -> str:
+        return TRAINER_NAME
