@@ -2,31 +2,32 @@
 # Contains an implementation of SAC as described in https://arxiv.org/abs/1801.01290
 # and implemented in https://github.com/hill-a/stable-baselines
 
-from collections import defaultdict
-from typing import Dict, cast
-import os
+from typing import cast
 
 import numpy as np
-from mlagents.trainers.policy.checkpoint_manager import ModelCheckpoint
 
 from mlagents_envs.logging_util import get_logger
-from mlagents_envs.timers import timed
 from mlagents_envs.base_env import BehaviorSpec
-from mlagents.trainers.buffer import BufferKey, RewardSignalUtil
-from mlagents.trainers.policy import Policy
-from mlagents.trainers.trainer.rl_trainer import RLTrainer
+from mlagents.trainers.buffer import BufferKey
+from mlagents.trainers.optimizer.torch_optimizer import TorchOptimizer
+from mlagents.trainers.trainer.off_policy_trainer import OffPolicyTrainer
 from mlagents.trainers.policy.torch_policy import TorchPolicy
-from mlagents.trainers.sac.optimizer_torch import TorchSACOptimizer
+from mlagents.trainers.policy.policy import Policy
+from mlagents.trainers.sac.optimizer_torch import TorchSACOptimizer, SACSettings
 from mlagents.trainers.trajectory import Trajectory, ObsUtil
 from mlagents.trainers.behavior_id_utils import BehaviorIdentifiers
-from mlagents.trainers.settings import TrainerSettings, SACSettings
+from mlagents.trainers.settings import TrainerSettings
+
+from mlagents.trainers.torch_entities.networks import SimpleActor
 
 logger = get_logger(__name__)
 
 BUFFER_TRUNCATE_PERCENT = 0.8
 
+TRAINER_NAME = "sac"
 
-class SACTrainer(RLTrainer):
+
+class SACTrainer(OffPolicyTrainer):
     """
     The SACTrainer is an implementation of the SAC algorithm, with support
     for discrete actions and recurrent networks.
@@ -54,15 +55,16 @@ class SACTrainer(RLTrainer):
         """
         super().__init__(
             behavior_name,
+            reward_buff_cap,
             trainer_settings,
             training,
             load,
+            seed,
             artifact_path,
-            reward_buff_cap,
         )
 
         self.seed = seed
-        self.policy: Policy = None  # type: ignore
+        self.policy: TorchPolicy = None  # type: ignore
         self.optimizer: TorchSACOptimizer = None  # type: ignore
         self.hyperparameters: SACSettings = cast(
             SACSettings, trainer_settings.hyperparameters
@@ -80,51 +82,6 @@ class SACTrainer(RLTrainer):
 
         self.checkpoint_replay_buffer = self.hyperparameters.save_replay_buffer
 
-    def _checkpoint(self) -> ModelCheckpoint:
-        """
-        Writes a checkpoint model to memory
-        Overrides the default to save the replay buffer.
-        """
-        ckpt = super()._checkpoint()
-        if self.checkpoint_replay_buffer:
-            self.save_replay_buffer()
-        return ckpt
-
-    def save_model(self) -> None:
-        """
-        Saves the final training model to memory
-        Overrides the default to save the replay buffer.
-        """
-        super().save_model()
-        if self.checkpoint_replay_buffer:
-            self.save_replay_buffer()
-
-    def save_replay_buffer(self) -> None:
-        """
-        Save the training buffer's update buffer to a pickle file.
-        """
-        filename = os.path.join(self.artifact_path, "last_replay_buffer.hdf5")
-        logger.info(f"Saving Experience Replay Buffer to {filename}...")
-        with open(filename, "wb") as file_object:
-            self.update_buffer.save_to_file(file_object)
-            logger.info(
-                f"Saved Experience Replay Buffer ({os.path.getsize(filename)} bytes)."
-            )
-
-    def load_replay_buffer(self) -> None:
-        """
-        Loads the last saved replay buffer from a file.
-        """
-        filename = os.path.join(self.artifact_path, "last_replay_buffer.hdf5")
-        logger.info(f"Loading Experience Replay Buffer from {filename}...")
-        with open(filename, "rb+") as file_object:
-            self.update_buffer.load_from_file(file_object)
-        logger.debug(
-            "Experience replay buffer has {} experiences.".format(
-                self.update_buffer.num_experiences
-            )
-        )
-
     def _process_trajectory(self, trajectory: Trajectory) -> None:
         """
         Takes a trajectory and processes it, putting it into the replay buffer.
@@ -139,7 +96,7 @@ class SACTrainer(RLTrainer):
 
         # Update the normalization
         if self.is_training:
-            self.policy.update_normalization(agent_buffer_trajectory)
+            self.policy.actor.update_normalization(agent_buffer_trajectory)
             self.optimizer.critic.update_normalization(agent_buffer_trajectory)
 
         # Evaluate all reward functions for reporting purposes
@@ -184,44 +141,12 @@ class SACTrainer(RLTrainer):
         if trajectory.done_reached:
             self._update_end_episode_stats(agent_id, self.optimizer)
 
-    def _is_ready_update(self) -> bool:
-        """
-        Returns whether or not the trainer has enough elements to run update model
-        :return: A boolean corresponding to whether or not _update_policy() can be run
-        """
-        return (
-            self.update_buffer.num_experiences >= self.hyperparameters.batch_size
-            and self._step >= self.hyperparameters.buffer_init_steps
-        )
+    def create_optimizer(self) -> TorchOptimizer:
+        return TorchSACOptimizer(  # type: ignore
+            cast(TorchPolicy, self.policy), self.trainer_settings  # type: ignore
+        )  # type: ignore
 
-    @timed
-    def _update_policy(self) -> bool:
-        """
-        Update the SAC policy and reward signals. The reward signal generators are updated using different mini batches.
-        By default we imitate http://arxiv.org/abs/1809.02925 and similar papers, where the policy is updated
-        N times, then the reward signals are updated N times.
-        :return: Whether or not the policy was updated.
-        """
-        policy_was_updated = self._update_sac_policy()
-        self._update_reward_signals()
-        return policy_was_updated
-
-    def maybe_load_replay_buffer(self):
-        # Load the replay buffer if load
-        if self.load and self.checkpoint_replay_buffer:
-            try:
-                self.load_replay_buffer()
-            except (AttributeError, FileNotFoundError):
-                logger.warning(
-                    "Replay buffer was unable to load, starting from scratch."
-                )
-            logger.debug(
-                "Loaded update buffer with {} sequences".format(
-                    self.update_buffer.num_experiences
-                )
-            )
-
-    def create_torch_policy(
+    def create_policy(
         self, parsed_behavior_id: BehaviorIdentifiers, behavior_spec: BehaviorSpec
     ) -> TorchPolicy:
         """
@@ -230,140 +155,18 @@ class SACTrainer(RLTrainer):
         :param behavior_spec: specifications for policy construction
         :return policy
         """
+        actor_cls = SimpleActor
+        actor_kwargs = {"conditional_sigma": True, "tanh_squash": True}
+
         policy = TorchPolicy(
             self.seed,
             behavior_spec,
-            self.trainer_settings,
-            condition_sigma_on_obs=True,
-            tanh_squash=True,
-            separate_critic=True,
+            self.trainer_settings.network_settings,
+            actor_cls,
+            actor_kwargs,
         )
         self.maybe_load_replay_buffer()
         return policy
-
-    def _update_sac_policy(self) -> bool:
-        """
-        Uses update_buffer to update the policy. We sample the update_buffer and update
-        until the steps_per_update ratio is met.
-        """
-        has_updated = False
-        self.cumulative_returns_since_policy_update.clear()
-        n_sequences = max(
-            int(self.hyperparameters.batch_size / self.policy.sequence_length), 1
-        )
-
-        batch_update_stats: Dict[str, list] = defaultdict(list)
-        while (
-            self._step - self.hyperparameters.buffer_init_steps
-        ) / self.update_steps > self.steps_per_update:
-            logger.debug(f"Updating SAC policy at step {self._step}")
-            buffer = self.update_buffer
-            if self.update_buffer.num_experiences >= self.hyperparameters.batch_size:
-                sampled_minibatch = buffer.sample_mini_batch(
-                    self.hyperparameters.batch_size,
-                    sequence_length=self.policy.sequence_length,
-                )
-                # Get rewards for each reward
-                for name, signal in self.optimizer.reward_signals.items():
-                    sampled_minibatch[RewardSignalUtil.rewards_key(name)] = (
-                        signal.evaluate(sampled_minibatch) * signal.strength
-                    )
-
-                update_stats = self.optimizer.update(sampled_minibatch, n_sequences)
-                for stat_name, value in update_stats.items():
-                    batch_update_stats[stat_name].append(value)
-
-                self.update_steps += 1
-
-                for stat, stat_list in batch_update_stats.items():
-                    self._stats_reporter.add_stat(stat, np.mean(stat_list))
-                has_updated = True
-
-            if self.optimizer.bc_module:
-                update_stats = self.optimizer.bc_module.update()
-                for stat, val in update_stats.items():
-                    self._stats_reporter.add_stat(stat, val)
-
-        # Truncate update buffer if neccessary. Truncate more than we need to to avoid truncating
-        # a large buffer at each update.
-        if self.update_buffer.num_experiences > self.hyperparameters.buffer_size:
-            self.update_buffer.truncate(
-                int(self.hyperparameters.buffer_size * BUFFER_TRUNCATE_PERCENT)
-            )
-        return has_updated
-
-    def _update_reward_signals(self) -> None:
-        """
-        Iterate through the reward signals and update them. Unlike in PPO,
-        do it separate from the policy so that it can be done at a different
-        interval.
-        This function should only be used to simulate
-        http://arxiv.org/abs/1809.02925 and similar papers, where the policy is updated
-        N times, then the reward signals are updated N times. Normally, the reward signal
-        and policy are updated in parallel.
-        """
-        buffer = self.update_buffer
-        n_sequences = max(
-            int(self.hyperparameters.batch_size / self.policy.sequence_length), 1
-        )
-        batch_update_stats: Dict[str, list] = defaultdict(list)
-        while (
-            self._step - self.hyperparameters.buffer_init_steps
-        ) / self.reward_signal_update_steps > self.reward_signal_steps_per_update:
-            # Get minibatches for reward signal update if needed
-            reward_signal_minibatches = {}
-            for name in self.optimizer.reward_signals.keys():
-                logger.debug(f"Updating {name} at step {self._step}")
-                if name != "extrinsic":
-                    reward_signal_minibatches[name] = buffer.sample_mini_batch(
-                        self.hyperparameters.batch_size,
-                        sequence_length=self.policy.sequence_length,
-                    )
-            update_stats = self.optimizer.update_reward_signals(
-                reward_signal_minibatches, n_sequences
-            )
-            for stat_name, value in update_stats.items():
-                batch_update_stats[stat_name].append(value)
-            self.reward_signal_update_steps += 1
-
-            for stat, stat_list in batch_update_stats.items():
-                self._stats_reporter.add_stat(stat, np.mean(stat_list))
-
-    def create_sac_optimizer(self) -> TorchSACOptimizer:
-        return TorchSACOptimizer(  # type: ignore
-            cast(TorchPolicy, self.policy), self.trainer_settings  # type: ignore
-        )  # type: ignore
-
-    def add_policy(
-        self, parsed_behavior_id: BehaviorIdentifiers, policy: Policy
-    ) -> None:
-        """
-        Adds policy to trainer.
-        """
-        if self.policy:
-            logger.warning(
-                "Your environment contains multiple teams, but {} doesn't support adversarial games. Enable self-play to \
-                    train adversarial games.".format(
-                    self.__class__.__name__
-                )
-            )
-        self.policy = policy
-        self.policies[parsed_behavior_id.behavior_id] = policy
-        self.optimizer = self.create_sac_optimizer()
-        for _reward_signal in self.optimizer.reward_signals.keys():
-            self.collected_rewards[_reward_signal] = defaultdict(lambda: 0)
-
-        self.model_saver.register(self.policy)
-        self.model_saver.register(self.optimizer)
-        self.model_saver.initialize_or_load()
-
-        # Needed to resume loads properly
-        self._step = policy.get_current_step()
-        # Assume steps were updated at the correct ratio before
-        self.update_steps = int(max(1, self._step / self.steps_per_update))
-        self.reward_signal_update_steps = int(
-            max(1, self._step / self.reward_signal_steps_per_update)
-        )
 
     def get_policy(self, name_behavior_id: str) -> Policy:
         """
@@ -372,3 +175,7 @@ class SACTrainer(RLTrainer):
         """
 
         return self.policy
+
+    @staticmethod
+    def get_trainer_name() -> str:
+        return TRAINER_NAME
