@@ -10,8 +10,6 @@ from mlagents.trainers.settings import (
     RunOptions,
     TrainerSettings,
     NetworkSettings,
-    PPOSettings,
-    SACSettings,
     RewardSignalType,
     RewardSignalSettings,
     CuriositySettings,
@@ -21,19 +19,25 @@ from mlagents.trainers.settings import (
     UniformSettings,
     GaussianSettings,
     MultiRangeUniformSettings,
-    TrainerType,
     deep_update_dict,
     strict_to_cls,
     ScheduleType,
 )
+from mlagents.trainers.ppo.trainer import PPOSettings, TRAINER_NAME as PPO_TRAINER_NAME
+from mlagents.trainers.sac.trainer import SACSettings, TRAINER_NAME as SAC_TRAINER_NAME
+
 from mlagents.trainers.exception import TrainerConfigError
+
+TRAINER_SETTING_TYPES = {"ppo": PPOSettings, "sac": SACSettings}
 
 
 def check_if_different(testobj1: object, testobj2: object) -> None:
     assert testobj1 is not testobj2
     if attr.has(testobj1.__class__) and attr.has(testobj2.__class__):
         for key, val in attr.asdict(testobj1, recurse=False).items():
-            if isinstance(val, dict) or isinstance(val, list) or attr.has(val):
+            if (
+                isinstance(val, dict) or isinstance(val, list) or attr.has(val)
+            ) and val != {}:
                 # Note: this check doesn't check the contents of mutables.
                 check_if_different(val, attr.asdict(testobj2, recurse=False)[key])
 
@@ -121,19 +125,20 @@ def test_trainersettings_structure():
     Test structuring method for TrainerSettings
     """
     trainersettings_dict = {
-        "trainer_type": "sac",
+        "trainer_type": SAC_TRAINER_NAME,
         "hyperparameters": {"batch_size": 1024},
         "max_steps": 1.0,
         "reward_signals": {"curiosity": {"encoding_size": 64}},
     }
     trainer_settings = TrainerSettings.structure(trainersettings_dict, TrainerSettings)
+    # check_trainer_setting_types([trainer_settings], TRAINER_SETTING_TYPES)
     assert isinstance(trainer_settings.hyperparameters, SACSettings)
-    assert trainer_settings.trainer_type == TrainerType.SAC
+    assert trainer_settings.trainer_type == SAC_TRAINER_NAME
     assert isinstance(trainer_settings.max_steps, int)
     assert RewardSignalType.CURIOSITY in trainer_settings.reward_signals
 
     # Check invalid trainer type
-    with pytest.raises(ValueError):
+    with pytest.raises(TrainerConfigError):
         trainersettings_dict = {
             "trainer_type": "puppo",
             "hyperparameters": {"batch_size": 1024},
@@ -144,7 +149,7 @@ def test_trainersettings_structure():
     # Check invalid hyperparameter
     with pytest.raises(TrainerConfigError):
         trainersettings_dict = {
-            "trainer_type": "ppo",
+            "trainer_type": PPO_TRAINER_NAME,
             "hyperparameters": {"notahyperparam": 1024},
             "max_steps": 1.0,
         }
@@ -166,7 +171,7 @@ def test_trainersettingsschedules_structure():
     Test structuring method for Trainer Settings Schedule
     """
     trainersettings_dict = {
-        "trainer_type": "ppo",
+        "trainer_type": PPO_TRAINER_NAME,
         "hyperparameters": {
             "learning_rate_schedule": "linear",
             "beta_schedule": "constant",
@@ -550,7 +555,9 @@ def test_default_settings():
 
     # Change the overridden fields back, and check if the rest are equal.
     test1_settings.max_steps = 1
-    test1_settings.network_settings.hidden_units == default_settings_cls.network_settings.hidden_units
+    test1_settings.network_settings.hidden_units = (
+        default_settings_cls.network_settings.hidden_units
+    )
     check_if_different(test1_settings, default_settings_cls)
 
 
